requirements.txt
Cython>=0.25.2
numpy>=1.16.3
scikit-image>=0.12.3
scipy>=0.18.1
tensorflow-gpu==1.14.0
opencv-python>=3.4.2.17
Pillow==4.0.0